Change the behavior of the new proposal link when not signed in
In case the privacy module is installed, the user was previously
redirected to the new proposal page showing the profile publicity
note when they used an omniauth sign in method.

This changes the behavior so that the user is redirected back to
the proposals page and the profile publicity modal is opened
instead.