Make map tooltips prettier
Map tooltips had some issues with background color and random borders around the tooltip. The tooltip has been customized now to fit the light theme of the decidim instances and description has been added to the tooltip with a 2-row limit for the body text. The changes on this commit have been made only to proposals and meetings, so if needed on other components they still need customizing